AuthForge-Auto/requirements.txt
fastapi==0.100.0
pytest==7.4.0
PyJWT==2.8.0
httpx==0.24.1
uvicorn==0.23.1
pydantic==2.1.1
pytest-cov==4.1.0
